Most updated schema for the kinases database
This is the most updated and probably the final schema for the kinases database, which is empty.  

It consists of all the templates for the tables saved as classes. These templates for the tables in this script are used to identify the featutres of the table (number of columns, datatype , length, special characters, relationships) and lead SQLalchemy to find the info and must be included in the scripts for the data population with csv files and the queries used in Flask.