kinases database with data
The transfer of data was done using SQL Studio as a quality control for the schema and the data from the csv files. 
The final database needs to be populated properly with csv files and python scripts though.
In the meantime, please use it to try queries on Flask but consider that it may have potential errors.